clean up topups page
